requirements.txt
gym==0.26.2
matplotlib==3.8.0
numpy==1.26.0
opencv_python==4.8.1.78
pandas==2.1.1
Pillow==9.0.1
Pillow==10.0.1
pycocotools==2.0.7
skimage==0.0
stable_baselines3==2.1.0
tensorboardX==2.6.2.2
torch==2.0.1
torchvision==0.16.0
tqdm==4.66.1
